Initial commit; fully working DNS and SPF parser modules
The modules are complete for this initial check-in.
This is enough for anyone else to get started.

The SPF is mostly RFC-copmliant, even for the lesser-used functions,
although is still missing the ptr type. To be added later.

Future additions will add a wrapper that _uses_ the module output
in some way, but this is most of the core code to parse SPF.

Testing has been limited at this early stage.