Bugfixes and PTR typre removal
The MX record type didn't correctly deal with the array/List of A records;
it does now. This bug has been fixed.

The PTR type wasa run-time check only, so I removed it,
as it makes no sense ahead of time. Also, PTR records are discouraged and
deprecated, and hardly anyone uses them, so this is no major loss.